WIP Major update - Removed nuklear - Replaced Test_Suite with sandbox - Migrating from SDL2 to SDL3

+++ SWGTK/engine/include/TTFFont.hpp
@@ -4,12 +4,11 @@
 #include <filesystem>
 #include <unordered_map>
 #include <cstdint>
-#include "SDL_ttf.h"
+#include "SDL3_ttf/SDL_ttf.h"
 
 namespace swgtk::sdl
 {
-    enum class FontStyle : int32_t
-    {
+    enum class FontStyle : int32_t {
         None = 32,
         Normal = TTF_STYLE_NORMAL,
         Bold = TTF_STYLE_BOLD,
@@ -25,8 +24,7 @@ namespace swgtk::sdl
         Italic_Strike = Italic | Strikethrough,
     };
 
-    class FontGroup
-    {
+    class FontGroup {
     public:
         FontGroup() = default;
         ~FontGroup() = default;
@@ -36,15 +34,14 @@ namespace swgtk::sdl
         FontGroup& operator=(const FontGroup&) = delete;
         FontGroup& operator=(FontGroup&&) = delete;
 
-        void AddFont(FontStyle styleMask, int size, const std::filesystem::path& filename);
+        void AddFont(FontStyle styleMask, float size, const std::filesystem::path& filename);
         void ClearTTFFonts();
 
         [[nodiscard]] TTF_Font* GetTTF(FontStyle mask, int size);
         [[nodiscard]] const TTF_Font* GetTTF(FontStyle mask, int size) const;
 
     private:
-        [[nodiscard]] constexpr static int64_t Hash(FontStyle style, int size)
-        {
+        [[nodiscard]] constexpr static int64_t Hash(FontStyle style, int size) {
             constexpr const int styleOffset = 32;
             return (int64_t)style << styleOffset | (int64_t)size;
         }
